requirements.txt
opencv-contrib-python==4.4.0.46
numpy==1.21.6
tqdm==4.64.0
pyyaml